#322 [refactor]: info/Members.tsx 서버컴포넌트로 만들기
